+++ front-end/uibuilder.iife.js
@@ -7185,6 +7185,10 @@
       // TODO Move to proper getters
       // #region ---- Externally read-only (via .get method) ---- //
       // version - moved to _meta
+      // Has the initial Socket.IO connection been made? Used to detect if a page reload or just a disconnect/reconnect
+      __publicField(this, "initialConnect", null);
+      // How many times has the client reconnected to Socket.IO since the page was loaded? Note that this will be 0 on the initial connection, so can be used to detect a page reload vs a disconnect/reconnect
+      __publicField(this, "reconnected", 0);
       /** Client ID set by uibuilder on connect */
       __publicField(this, "clientId", "");
       /** The collection of cookies provided by uibuilder */
@@ -9788,7 +9792,7 @@
       this.set("ioConnected", false);
       if (__privateGet(this, _timerid)) window.clearTimeout(__privateGet(this, _timerid));
       __privateSet(this, _timerid, window.setTimeout(() => {
-        log("warn", "Uib:checkConnect:setTimeout", "Socket.IO reconnection attempt. Current delay: ".concat(delay, ". Depth: ").concat(depth))();
+        log("warn", "Uib:checkConnect:setTimeout", "Socket.IO reconnection attempt. Current delay: ".concat(delay, ". Depth: ").concat(depth, ". Initial Connect: ").concat(this.initialConnect, ". Reconnected: ").concat(this.reconnected))();
         this._socket.disconnect();
         this._socket.connect();
         __privateSet(this, _timerid, null);
@@ -9803,10 +9807,12 @@
      */
     _onConnect() {
       this.currentTransport = this._socket.io.engine.transport.name;
+      if (this.initialConnect === null) this.initialConnect = true;
+      else this.reconnected++;
       log(
         "info",
         "Uib:ioSetup",
-        "\u2705 SOCKET CONNECTED.\nConnection count: ".concat(this.connectedNum, ", Is a Recovery?: ").concat(this._socket.recovered, ".\nNamespace: ").concat(this.ioNamespace, "\nTransport: ").concat(this.currentTransport)
+        "\u2705 SOCKET CONNECTED.\nConnection count: ".concat(this.connectedNum, " (Reconnected: ").concat(this.reconnected, ", Initial Connect: ").concat(this.initialConnect, "), Is a Recovery?: ").concat(this._socket.recovered, ".\nNamespace: ").concat(this.ioNamespace, "\nTransport: ").concat(this.currentTransport)
       )();
       this._dispatchCustomEvent("uibuilder:socket:connected", { numConnections: this.connectedNum, isRecovery: this._socket.recovered });
       this._socket.io.engine.on("upgrade", () => {
@@ -9814,7 +9820,7 @@
         log(
           "trace",
           "Uib:_onConnect:onUpgrade",
-          "SOCKET CONNECTION UPGRADED.\nConnection count: ".concat(this.connectedNum, ", Is a Recovery?: ").concat(this._socket.recovered, ".\nNamespace: ").concat(this.ioNamespace, "\nTransport: ").concat(this.currentTransport)
+          "SOCKET CONNECTION UPGRADED.\nConnection count: ".concat(this.connectedNum, " (Reconnected: ").concat(this.reconnected, ", Initial Connect: ").concat(this.initialConnect, "), Is a Recovery?: ").concat(this._socket.recovered, ".\nNamespace: ").concat(this.ioNamespace, "\nTransport: ").concat(this.currentTransport)
         )();
       });
       setTimeout(() => {
@@ -9822,7 +9828,7 @@
           log(
             "error",
             "Uib:Connection",
-            "Connected to Node-RED but NO SOCKET UPGRADE!\n\u27A1\uFE0F CHECK NETWORK and any PROXIES for issues. \u2B05\uFE0F\nConnection count: ".concat(this.connectedNum, ", Is a Recovery?: ").concat(this._socket.recovered, ".\nNamespace: ").concat(this.ioNamespace, "\nTransport: ").concat(this.currentTransport)
+            "Connected to Node-RED but NO SOCKET UPGRADE!\n\u27A1\uFE0F CHECK NETWORK and any PROXIES for issues. \u2B05\uFE0F\nConnection count: ".concat(this.connectedNum, " (Reconnected: ").concat(this.reconnected, ", Initial Connect: ").concat(this.initialConnect, "), Is a Recovery?: ").concat(this._socket.recovered, ".\nNamespace: ").concat(this.ioNamespace, "\nTransport: ").concat(this.currentTransport)
           )();
         }
       }, 2e3);